Check that requirements are actually on interfaces
