+++ src/tests/stmt/funcion_non_interface_requirement1.hay
@@ -1,0 +1,2 @@
+fn Foo<T>() 
+requires: [Str] { }
