+++ src/tests/stmt/funcion_non_interface_requirement2.hay
@@ -1,0 +1,2 @@
+fn Foo<T>()
+requires: [Arr<T>] { }
