+++ src/tests/stmt/interface_non_interface_requirement2.hay
@@ -1,0 +1,2 @@
+interface Foo<T>
+requires: [Arr<T>] { }
