+++ src/tests/stmt/record_non_interface_requirement1.hay
@@ -1,0 +1,4 @@
+struct Foo<T>
+requires: [Str] {
+    T: t
+}
