+++ src/tests/stmt/record_non_interface_requirement2.hay
@@ -1,0 +1,4 @@
+union Foo<T>
+requires: [Arr<T>] {
+    T: t
+}
